# bank_rest/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bank Cards Management API
  description: |
    REST API для системы управления банковскими картами и переводами.
    
    ## Основные возможности:
    - Управление пользователями и аутентификация
    - Создание и управление банковскими картами
    - Выполнение денежных переводов и транзакций
    - Безопасность на основе JWT токенов
    - Маскирование конфиденциальных данных
    
    ## Роли и доступ:
    - **ADMIN**: Полный доступ ко всем операциям
    - **USER**: Доступ только к своим картам и операциям
    
    ## Авторизация:
    Для доступа к защищенным endpoints необходимо:
    1. Вызвать `/api/auth/login` для получения JWT токена
    2. Добавить заголовок: `Authorization: Bearer <your_token>`
  version: 1.0.0
  contact:
    name: Support Team
    email: support@bankapi.com
    url: https://bankapi.com/support
  license:
    name: Proprietary
    url: https://bankapi.com/license
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.bankcards.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите JWT токен полученный при аутентификации
  schemas:
...
    CardType:
      type: string
      enum: [DEBIT, CREDIT]
      description: Тип банковской карты
    Currency:
      type: string
      enum: [USD, RUB]
      description: Валюта карты
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Успешность операции
        message:
          type: string
          description: Сообщение о результате
        data:
          type: object
          description: Данные ответа
        timestamp:
          type: string
          format: date-time
          description: Время ответа
        path:
          type: string
          description: Путь запроса
    PageResponse:
      type: object
      properties:
        content:
          type: array
          description: Список элементов на текущей странице
        currentPage:
          type: integer
          example: 0
          description: Номер текущей страницы
        pageSize:
          type: integer
          example: 20
          description: Размер страницы
        totalElements:
          type: integer
          example: 150
          description: Общее количество элементов
        totalPages:
          type: integer
          example: 8
          description: Общее количество страниц
        hasNext:
          type: boolean
          example: true
          description: Есть ли следующая страница
        hasPrevious:
          type: boolean
          example: false
          description: Есть ли предыдущая страница
...
    CardDTO:
      type: object
      properties:
        id:
          type: integer
          description: ID карты
        cardNumber:
          type: string
          example: "**** **** **** 1234"
          description: Номер карты (маскированный)
        cardHolderName:
          type: string
          example: "IVAN IVANOV"
          description: Имя владельца карты
        expiryDate:
          type: string
          format: date
          example: "2025-12-31"
          description: Дата истечения срока действия
        cardType:
          $ref: '#/components/schemas/CardType'
        currency:
          $ref: '#/components/schemas/Currency'
        balance:
          type: number
          format: double
          example: 1000.50
          description: Баланс карты
        creditLimit:
          type: number
          format: double
          example: 50000.00
          description: Кредитный лимит
        availableCredit:
          type: number
          format: double
          example: 45000.00
          description: Доступный кредит
        active:
          type: boolean
          description: Активна ли карта
        isBlocked:
          type: boolean
          description: Заблокирована ли карта
        createdAt:
          type: string
          format: date-time
          description: Дата создания карты
        userId:
          type: integer
          description: ID владельца карты
        userFullName:
          type: string
          description: Полное имя владельца
    CreateCardRequest:
      type: object
      required:
        - cardType
        - currency
        - userId
      properties:
        cardType:
          $ref: '#/components/schemas/CardType'
        currency:
          $ref: '#/components/schemas/Currency'
        userId:
          type: integer
          example: 1
          description: ID пользователя для привязки карты
    CardStatusResponse:
      type: object
      properties:
        status:
          type: string
          example: "ACTIVE"
          enum: [ACTIVE, BLOCKED, EXPIRED, DEACTIVATED]
          description: Статус карты
        statusDescription:
          type: string
          description: Детальное описание статуса
        active:
          type: boolean
          description: Активна ли карта
        blocked:
          type: boolean
          description: Заблокирована ли карта
        expired:
          type: boolean
          description: Истек ли срок действия
        valid:
          type: boolean
          description: Действительна ли карта
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
        daysUntilExpiry:
          type: integer
          description: Дней до истечения срока действия
        card:
          $ref: '#/components/schemas/CardDTO'
    CardSearchRequest:
      type: object
      properties:
        cardNumber:
          type: string
          example: "1234"
          description: Часть номера карты для поиска
        cardType:
          $ref: '#/components/schemas/CardType'
        currency:
          $ref: '#/components/schemas/Currency'
        active:
          type: boolean
          description: Статус активности карты
        isBlocked:
          type: boolean
          description: Статус блокировки карты
        page:
          type: integer
          minimum: 0
          default: 0
          example: 0
          description: Номер страницы (начинается с 0)
        size:
          type: integer
          minimum: 1
          default: 20
          example: 20
          description: Размер страницы
        sort:
          type: string
          default: "createdAt,desc"
          example: "createdAt,desc"
          description: "Поле для сортировки (формат: field,asc|desc)"
...
security:
  - bearerAuth: []
paths:
...
  /api/cards:
    get:
      tags: [Cards]
      summary: Получить список карт
      description: |
        Получение списка карт с пагинацией.
        - **USER**: только свои карты
        - **ADMIN**: все карты
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Номер страницы
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            default: "createdAt,desc"
          description: Сортировка
      responses:
        '200':
          description: Список карт получен успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '403':
          description: Доступ запрещен

    post:
      tags: [Cards]
      summary: Создать новую карту (ADMIN)
      description: Создание новой банковской карты для пользователя
      security:
        - bearerAuth: [ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '400':
          description: Ошибка валидации
        '403':
          description: Недостаточно прав
        '404':
          description: Пользователь не найден
  /api/cards/search:
    post:
      tags: [Cards]
      summary: Поиск карт с фильтрацией
      description: |
        Поиск карт с расширенной фильтрацией и пагинацией.
        - **USER**: поиск только по своим картам
        - **ADMIN**: поиск по всем картам
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardSearchRequest'
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '403':
          description: Доступ запрещен
  /api/cards/{id}:
    get:
      tags: [Cards]
      summary: Получить карту по ID
      description: Получение детальной информации о карте
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '403':
          description: Доступ запрещен (USER пытается получить не свою карту)
        '404':
          description: Карта не найдена

    delete:
      tags: [Cards]
      summary: Удалить карту (ADMIN)
      description: Удаление карты из системы
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Карта успешно удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Недостаточно прав
        '404':
          description: Карта не найдена
  /api/cards/{id}/status:
    get:
      tags: [Cards]
      summary: Получить статус карты
      description: Получение детальной информации о статусе карты
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Статус карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardStatusResponse'
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
  /api/cards/user/{userId}:
    get:
      tags: [Cards]
      summary: Получить карты пользователя
      description: |
        Получение списка карт конкретного пользователя.
        - **USER**: только свои карты
        - **ADMIN**: карты любого пользователя
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID пользователя
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Номер страницы
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Размер страницы
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '403':
          description: Доступ запрещен
  /api/cards/{id}/block:
    put:
      tags: [Cards]
      summary: Заблокировать карту
      description: |
        Блокировка карты.
        - **USER**: может блокировать только свои карты
        - **ADMIN**: может блокировать любые карты
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Карта успешно заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
  /api/cards/{id}/unblock:
    put:
      tags: [Cards]
      summary: Разблокировать карту (ADMIN)
      description: Разблокировка карты администратором
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Карта успешно разблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Недостаточно прав
        '404':
          description: Карта не найдена
  /api/cards/{id}/balance:
    get:
      tags: [Cards]
      summary: Получить баланс карты
      description: Получение текущего баланса карты
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Баланс карты
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    format: double
                  currency:
                    $ref: '#/components/schemas/Currency'
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
  /api/cards/total-balance:
    get:
      tags: [Cards]
      summary: Получить общий баланс
      description: Получение общего баланса по всем картам пользователя
      responses:
        '200':
          description: Общий баланс
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalBalance:
                    type: number
                    format: double
                  cardsCount:
                    type: integer
                  currencies:
                    type: array
                    items:
                      type: object
                      properties:
                        currency:
                          $ref: '#/components/schemas/Currency'
                        balance:
                          type: number
                          format: double
        '403':
          description: Доступ запрещен
  /api/cards/{id}/validate:
    get:
      tags: [Cards]
      summary: Проверить валидность карты
      description: Проверка действительности карты (активна, не заблокирована, не просрочена)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Действительна ли карта
                  reasons:
                    type: array
                    items:
                      type: string
                    description: Причины невалидности (если есть)
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена

  # Transactions endpoints
...